dev/requirements.txt
aiogram==3.3.0
aiohttp==3.9.5
openai==1.25.0
pillow==10.0.0            